execversion: extract DistSQL version into a new package
The following commits will introduce dependency on this version into
other low-dependency packages, so this extraction is needed to avoid
import cycles.

Release note: None
