+++ pkg/BUILD.bazel
@@ -467,6 +467,7 @@ ALL_TESTS = [
     "//pkg/sql/execinfrapb:execinfrapb_test",
     "//pkg/sql/execstats:execstats_disallowed_imports_test",
     "//pkg/sql/execstats:execstats_test",
+    "//pkg/sql/execversion:execversion_test",
     "//pkg/sql/exprutil:exprutil_test",
     "//pkg/sql/flowinfra:flowinfra_disallowed_imports_test",
     "//pkg/sql/flowinfra:flowinfra_test",
@@ -1937,6 +1938,8 @@ GO_TARGETS = [
     "//pkg/sql/execinfrapb:execinfrapb_test",
     "//pkg/sql/execstats:execstats",
     "//pkg/sql/execstats:execstats_test",
+    "//pkg/sql/execversion:execversion",
+    "//pkg/sql/execversion:execversion_test",
     "//pkg/sql/exprutil:evalexpr",
     "//pkg/sql/exprutil:exprutil",
     "//pkg/sql/exprutil:exprutil_test",
