+++ pkg/sql/BUILD.bazel
@@ -424,6 +424,7 @@ go_library(
         "//pkg/sql/execinfra/execreleasable",
         "//pkg/sql/execinfrapb",
         "//pkg/sql/execstats",
+        "//pkg/sql/execversion",
         "//pkg/sql/exprutil",
         "//pkg/sql/faketreeeval",
         "//pkg/sql/flowinfra",
