+++ pkg/sql/distsql/BUILD.bazel
@@ -17,6 +17,7 @@ go_library(
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfra/execopnode",
         "//pkg/sql/execinfrapb",
+        "//pkg/sql/execversion",
         "//pkg/sql/faketreeeval",
         "//pkg/sql/flowinfra",
         "//pkg/sql/rowflow",
@@ -81,6 +82,7 @@ go_test(
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfra/execagg",
         "//pkg/sql/execinfrapb",
+        "//pkg/sql/execversion",
         "//pkg/sql/flowinfra",
         "//pkg/sql/randgen",
         "//pkg/sql/rowenc",
