+++ pkg/sql/distsql/server.go
@@ -21,6 +21,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra/execopnode"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
+	"github.com/cockroachdb/cockroach/pkg/sql/execversion"
 	"github.com/cockroachdb/cockroach/pkg/sql/faketreeeval"
 	"github.com/cockroachdb/cockroach/pkg/sql/flowinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowflow"
@@ -207,10 +208,10 @@ func (ds *ServerImpl) setupFlow(
 		}
 	}()
 
-	if req.Version < execinfra.MinAcceptedVersion || req.Version > execinfra.Version {
+	if req.Version < execversion.MinAcceptedVersion || req.Version > execversion.Version {
 		err := errors.Errorf(
 			"version mismatch in flow request: %d; this node accepts %d through %d",
-			req.Version, execinfra.MinAcceptedVersion, execinfra.Version,
+			req.Version, execversion.MinAcceptedVersion, execversion.Version,
 		)
 		log.Warningf(ctx, "%v", err)
 		return ctx, nil, nil, err
