+++ pkg/sql/execinfrapb/api.go
@@ -42,9 +42,6 @@ func (f FlowID) IsUnset() bool {
 	return f.UUID.Equal(uuid.Nil)
 }
 
-// DistSQLVersion identifies DistSQL engine versions.
-type DistSQLVersion uint32
-
 // MakeEvalContext serializes some of the fields of a eval.Context into a
 // execinfrapb.EvalContext proto.
 func MakeEvalContext(evalCtx *eval.Context) EvalContext {
