+++ pkg/sql/execversion/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "execversion",
+    srcs = ["version.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/execversion",
+    visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "execversion_test",
+    srcs = ["version_test.go"],
+    embed = [":execversion"],
+    deps = [
+        "//pkg/util/leaktest",
+        "@com_github_stretchr_testify//require",
+    ],
+)
